+++ src/main/java/net/silentchaos512/gems/client/gui/DebugOverlay.java
@@ -27,6 +27,6 @@ public class DebugOverlay extends DebugRenderOverlay {
 
     @Override
     public boolean isHidden() {
-        return !(GemsConfig.COMMON.debugMasterSwitch.get() && GemsConfig.COMMON.debugShowOverlay.get());
+        return !GemsConfig.COMMON.debugShowOverlay.get();
     }
 }