+++ src/main/java/net/silentchaos512/gems/network/Network.java
@@ -4,13 +4,14 @@ import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.network.NetworkRegistry;
 import net.minecraftforge.fml.network.simple.SimpleChannel;
 import net.silentchaos512.gems.SilentGems;
+import net.silentchaos512.gems.lib.soul.Soul;
 
 import java.util.Objects;
 
 public final class Network {
     private static final ResourceLocation NAME = SilentGems.getId("network");
 
-    static SimpleChannel channel;
+    public static SimpleChannel channel;
     static {
         channel = NetworkRegistry.ChannelBuilder.named(NAME)
                 .clientAcceptedVersions(s -> Objects.equals(s, "1"))
@@ -22,6 +23,11 @@ public final class Network {
                 .decoder(GeneralSyncPacket::fromBytes)
                 .encoder(GeneralSyncPacket::toBytes)
                 .consumer(GeneralSyncPacket::handle)
+                .add();
+        channel.messageBuilder(SyncSoulsPacket.class, 2)
+                .decoder(SyncSoulsPacket::fromBytes)
+                .encoder(SyncSoulsPacket::toBytes)
+                .consumer(Soul::handleSyncPacket)
                 .add();
     }